update role user

+++ modules/backend-dossiermgt/backend-dossiermgt-service/src/main/java/org/opencps/dossiermgt/action/impl/ServiceProcessActionsImpl.java
@@ -258,7 +258,7 @@ public class ServiceProcessActionsImpl implements ServiceProcessActions {
 
 	@Override
 	public long updateServiceProcessDB(long userId, long groupId, String processNo, String processName,
-			String description, Integer durationCount, Integer durationUnit, boolean generatePassword, String serverNo,
+			String description, Double durationCount, Integer durationUnit, boolean generatePassword, String serverNo,
 			String serverName, String dossierNoPattern, String dueDatePattern, ServiceContext serviceContext)
 			throws PortalException {
